+++ react/__mocks__/vtex.shipping-option-components/ShippingOptionContext.ts
@@ -1,0 +1,3 @@
+export function useShippingOptionState() {
+  return { shippingOption: undefined }
+}